+++ test/inputs.test.ts
@@ -1,3 +1,5 @@
+import * as yaml from "js-yaml";
+
 import * as core from "./mocks/@actions/core.mock";
 
 jest.mock("@actions/core", () => core);
@@ -44,7 +46,7 @@ describe("ActionConfig::constructor", () => {
   });
 
   test("construct with an object specifying dry-run", () => {
-    const config: RawActionConfig = { "dry-run": "true" };
+    const config: RawActionConfig = { "dry-run": true };
     expect(() => new ActionConfig(config)).not.toThrow();
   });
 
@@ -108,7 +110,8 @@ describe("ActionConfig.isDryRun", () => {
   });
 
   test("dry-run is `false` in the config object", () => {
-    const instance: ActionConfig = new ActionConfig({ "dry-run": "false" });
+    const rawConfig: RawActionConfig = yaml.safeLoad("dry-run: false");
+    const instance: ActionConfig = new ActionConfig(rawConfig);
     core.getInput.mockReturnValueOnce("true");
 
     const result = instance.isDryRun();
@@ -116,7 +119,26 @@ describe("ActionConfig.isDryRun", () => {
   });
 
   test("dry-run is `true` in the config object", () => {
-    const instance: ActionConfig = new ActionConfig({ "dry-run": "true" });
+    const rawConfig: RawActionConfig = yaml.safeLoad("dry-run: true");
+    const instance: ActionConfig = new ActionConfig(rawConfig);
+    core.getInput.mockReturnValueOnce("false");
+
+    const result = instance.isDryRun();
+    expect(result).toBe(true);
+  });
+
+  test("dry-run is `'false'` in the config object", () => {
+    const rawConfig: RawActionConfig = yaml.safeLoad("dry-run: 'false'");
+    const instance: ActionConfig = new ActionConfig(rawConfig);
+    core.getInput.mockReturnValueOnce("true");
+
+    const result = instance.isDryRun();
+    expect(result).toBe(false);
+  });
+
+  test("dry-run is `'true'` in the config object", () => {
+    const rawConfig: RawActionConfig = yaml.safeLoad("dry-run: 'true'");
+    const instance: ActionConfig = new ActionConfig(rawConfig);
     core.getInput.mockReturnValueOnce("false");
 
     const result = instance.isDryRun();
@@ -124,7 +146,8 @@ describe("ActionConfig.isDryRun", () => {
   });
 
   testNonBoolean("dry run is `'%s'` in the config object", async (value) => {
-    const instance: ActionConfig = new ActionConfig({ "dry-run": value });
+    const rawConfig: RawActionConfig = yaml.safeLoad(`dry-run: '${value}'`);
+    const instance: ActionConfig = new ActionConfig(rawConfig);
     core.getInput.mockReturnValueOnce("false");
 
     const result = instance.isDryRun();