+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Versioning].
 
 - Configure the SVGO options file. ([#99])
 - Fix `required` value for "configuration-path" input. ([#100])
+- Fix interpretation of "dry-run" input in config file. ([#103])
 
 ## [0.2.1] - 2020-03-02
 
@@ -56,3 +57,4 @@ Versioning].
 [#97]: https://github.com/ericcornelissen/svgo-action/pull/97
 [#99]: https://github.com/ericcornelissen/svgo-action/pull/99
 [#100]: https://github.com/ericcornelissen/svgo-action/pull/100
+[#103]: https://github.com/ericcornelissen/svgo-action/pull/103